auth/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pymongo==4.10.1
PyJWT==2.9.0
passlib==1.7.4
bcrypt==4.1.3
grpcio==1.64.1
grpcio-tools==1.64.1
grpcio-reflection==1.64.1
protobuf==5.29.5
